Moved Apiresponse to a new file
