+++ package/api/__init__.py
@@ -1,1 +1,2 @@
-from ._api import YouTube, ApiResponse
+from ._api import YouTube
+from ._api_response import ApiResponse
